+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 WebConfigServer: Changelog
 ==========================
 
+HEAD
+----
+
+* Fix reconfigure NAPT after IP_EVENT_STA_GOT_IP for esp32
+
 v2.2.2 (2023-05-19)
 ------