+++ CHANGELOG.md
@@ -9,6 +9,7 @@ HEAD
 * Replace ESPAsyncWebServer fork dependency
 * Replace const JsonDocument with JsonVariantConst for config parse methods
 * Check if config json contains IWebConfig object or service before parsing it
+* Use ESPAsyncWebServer git repo as current version. Close issue #2
 
 v2.2.2 (2023-05-19)
 ------